app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
apscheduler==3.10.4
python-dateutil==2.9.0.post0
openpyxl==3.1.5
dnspython==2.6.1
aiofiles==23.2.1
